fix: concat api part of Kirby CMS url
Fix issue when using base URL of Kirby installation without the
API path. Previously, in opposite of what was documented in the readme,
the plugin assumed the full API query URL to be set in the environment
configuration. Now it except the base url of the Kirby installation,
as documented in the readme